relocate popover to certificates section
