+++ src/main/java/health/medunited/architecture/jaxrs/resource/Certificate.java
@@ -48,6 +48,7 @@ import de.gematik.ws.conn.eventservice.v7.GetCards;
 import de.gematik.ws.conn.eventservice.v7.GetCardsResponse;
 import de.gematik.ws.conn.eventservice.wsdl.v7.EventServicePortType;
 import de.gematik.ws.conn.eventservice.wsdl.v7.FaultMessage;
+import health.medunited.architecture.jaxrs.resource.crypto.CryptoLoader;
 import health.medunited.architecture.jaxrs.resource.model.VerifyAllEntry;
