+++ src/main/java/health/medunited/architecture/jaxrs/resource/crypto/IdpCryptoException.java
@@ -1,4 +1,4 @@
-package health.medunited.architecture.jaxrs.resource;
+package health.medunited.architecture.jaxrs.resource.crypto;
 
 public class IdpCryptoException extends Exception {
