+++ src/main/webapp/frontend/view/CertPopover.fragment.xml
@@ -3,7 +3,7 @@
 	xmlns:core="sap.ui.core">
 	<Popover
 		id="certPopover"
-		title="{Cards>cardType}"
+		title="{VerifyAll>certRef}"
 		class="sapUiResponsivePadding--header sapUiResponsivePadding--footer"
 		placement="Bottom"
 		initialFocus="popoverClose">
@@ -17,7 +17,7 @@
 			</OverflowToolbar>
 		</footer>
 		<content>
-			<Text text="{Cards>cardHandle}"/>
+			<Text text="{VerifyAll>cardHandle}"/>
 		</content>
 	</Popover>
 </core:FragmentDefinition>
